Further rewrite

+++ GoXLR-Utility.NET/EventArgs/Response/Status/Mixer/Levels/Volumes/SpecificVolumeChangedEventArgs.cs
@@ -4,6 +4,6 @@ namespace GoXLR_Utility.NET.EventArgs.Response.Status.Mixer.Levels.Volumes
     {
         public string SerialNumber { get; internal set; }
             
-        public int Volume { get; internal set; }
+        public byte Volume { get; internal set; }
     }
 }
